Remove redundand headers
